verify: use hash_raw_pod_slice to hash journal without trailer (#291)
Previously, the verifier used hash_words() to create a hash of the journal.
Unintentionally, this resulted in a "trailer" being appended to the hashed
journal when the input length was not a multiple of 16. In the guest, the
data that was committed gets hashed without a trailer if the data exceeds 32
bytes. The verifier needs the hash of the journal to match the seal root so the
divergent behavior between the verifier and guest causes unintended
verification errors for guest code that commits more than 32 bytes of data that
are not divisible by 16. Fix this by using a hashing function in the verifier
that does not add a trailer. Also, add a "hello world" test case as a
regression test.