[NFC] Refactor code in advance of adding support for alternate hashes (#367)
Modify the abstractions for Hal and VerifyHal so that they depend on 1) The field 2) The 'Hash Suite'. The hash suite in turn determines which hash function is used as well as how RNG in Fiat-Shamir is computed. This is done via two new traits 'ConfigHash' and 'ConfigRng' which provide the smallest possible API surface. These API's are implemented using the existing Sha256 based protocol when the HashSuiteSha256 is selected (and given a device appropriate Sha256 trait). In general, since there are currently no other HashSuite implementation, this entire PR is just a NFC code reorg.