Add BigInt syscall and accelerator in rv32im circuit (#466)
Many significant use cases for the RISC Zero zkVM require implementing finite field arithmetic in
fields of cryptographic sizes (e.g. 256 bits). A key example is verification of ECDSA signatures
over secp256k1. Currently, this is prohibitively expensive to do inside the zkVM.

In order to address this, we are adding a BigInt accelerator to the rv32im circuit. This PR contains
new compiled circuit as well as the accompanying Rust code to make the functionality accessible.

In the current implementation, only modular multiplication is supported. Each input is a 256-bit
number represented as an array of u32 values ordered from least to most significant. Using the new
sys_bigint call will result in the modular multiplication x * y == z mod N being written to the result pointer.

This PR is currently marked as a draft. In order make it ready to merge, the accelerator needs to
have its constraints optimized further. Additionally the interface code included in this PR needs to
be more fleshed out.

Following this PR, this functionality can be made available to developers by providing a patched
version of common cryptographic libraries, such as RustCrypto. A top candidate is to provide a
patched version of k256 using this new ecall to implement multiplication.