Add ECDSA verification example using bigint acceleration support (#590)
This PR adds an ECDSA verification example, which shows how to use the patched k256 library to get accelerated signature verification on secp256k1 in the zkVM guest.

As a dependency for the k256 patch, it also updates the allocator to support alignments that are larger than a word. (In particular, the constant-time lookup table implementation uses a 1024-byte alignment). risc0/rust#11 is related to this PR. Based on work by @shkoo.

As a drive-by change, this PR includes an additional check in the executor that execution did not previously end with ExitCode::Halted, which is a non-resumable state. I added this after running into this error by accident and having a hard time debugging it based on the errors I received.