Cleanup recursion structs (#632)
* ReceiptMeta: use ExitCode instead of two u8s

This is a first step in merging ReceiptMeta and ReceiptMetadata to a single
struct. The next change will remove ReceiptMeta in favor of ReceiptMetadata

* rename field in recursion::receipt::SystemState so it matches zkvm::receipt::SystemState

* recursion: remove duplicate `SystemState` struct

This struct was duplicated in the private repo. Since there is no need for
dupilcation, this struct has been removed.

* Recursion: remove duplicate ReceiptMeta struct

This change removes the duplicated ReceiptMetadata struct and instead uses a
common ReceiptMetadata struct in `risc0_zkvm::receipt`

* Move `exit_code` funtions to `risc0_zkvm::receipt`

These functions can be useful in `risc0_zkvm::receipt` as well as the recursion
module.

* clean up use declarations

* Fix documentation

---------

Co-authored-by: Frank Laub <flaub@risc0.com>