Bonsai Governance Showcase and Ethereum smart contract libraries. (#695)
This PR adds the Bonsai Governance Showcase to the new bonsai/examples directory. It also adds a set of Bonsai smart contract libraries, developed first as part of the Bonsai Foundry Template and significantly expanded as part of this work.

For more information about the Governance Showcase, see the attached README

After this PR is merged, I will also submit a PR to have the Bonsai Foundry Template import the smart contract libraries from this repository.