Reorg to support client/server architecture (#808)
## Why

The zkVM is moving towards a client/server architecture for two main
reasons:
1. Streamline the installation of the zkVM for most users.
2. Allow multiple versions of the zkVM (and hence circuits) to be used
by clients

The first is achieved by using prebuilt binaries that include HW
acceleration which is usually expensive to build from source. The second
is to rely on a stable interface between client and server, so that
newer clients can work with older servers. Both are supported by
decoupling the expensive zkVM bits from the user bits so that the
dependency graph of the client is minimized. It's also possible in the
future to support clients written in other languages.

## Reorg

The biggest change is a reorganization of the code:
* zkvm
  * guest
  * host
    * client
    * server
      * exec
      * prove

## Prover Traits

The `Prover` trait has been split into `Prover` and `DynProverImpl`.
There is now a distinction among client-side `Prover`s, and server-side
`ProverImpl`s. A `Prover` is a higher-level interface which includes
execution. A `ProverImpl` is only concerned with the proving phase.

A `ProverOpts` is added to the prove functions so that users can
configure the prover. Right now it only supports setting the hashing
function.

There are three client `Trait` impls:
* Bonsai
* External (calls `r0vm`, not yet complete)
* Local (only available when `prove` feature is enabled).

## Feature Flags

Currently the `prove` feature is enabled by default. In the future, once
the `ExternalProver` is complete and it's easy for users to install the
prebuilt `r0vm` via `cargo risczero install`, the `prove` feature will
not be enabled by default and should only be used by `r0vm`.

A new `client` feature is added to toggle support for the client-side.
The intention is for `default = ["client"]` once client/server is
complete. Pure verify (useful for WASM) is supported with
`default-features = false`.

## Changes

* `RISC0_DEV_MODE` now requires the value to be one of `["1", "true",
"yes"]` to be considered enabled.
* The `provers()` registry is dropped, along with `get_prover()`. Use
`get_prover_impl()` to select a `ProverImpl` based on `ProverOpts`.
* Drop support for low-level custom syscalls on the `ExecutorEnv`. It's
difficult and potentially dangerous to expose this across the
client/server interface.
* Dev mode is implemented as a `ProverImpl` rather than a `Prover` so
that it works independent of the hash function (or other `ProverOpts`)
that are selected.