Fault proof part 1 (#824)
This PR is step 1 in implementing proof of fault. When the rrs-lib crate runs into an instruction execution error, it triggers a guest program called the fault checker. The fault checker will take the previous state of execution and ensure that it will result in an error. This PR implements the plumbing for this along with the fault checker itself. As of this PR, the fault checker does not compute the POST ID as it should. I will follow up with a PR that implements the computation of the post image ID in the fault checker guest.

---------

Co-authored-by: Victor Graf <victor@risczero.com>