Add minimal Substrate WASM runtime for compatibility check (#911)
This PR aims to ensure ongoing compatibility between
[Substrate](https://substrate.io/) and RISC Zero through a test
including both. This is useful because:
- It serves as a check on `risc0-zkvm`'s compatibility in some no-std,
wasm environments
- It assures the RISC Zero verifier functions with Substrate blockchains

This change stemmed from a conversation with @SchmErik and myself in
Discord.

Changes:
- Inclusion of tiny Substrate runtime
- This example of a Substrate runtime has a minimal amount of Substrate
stuff necessary to be considered a runtime from a crates standpoint(it
does not actually function, and does not need to; it's only for
compatibility testing)
  - This runtime depends on the local `risc0-zkvm`
- CI command which tries to `cargo check` the Substrate runtime.

Notes:
- On future maintenance: Substrate is commonly versioned by
release-branch. This would need to be manually updated every quarter-1/2
year or so. Maybe there's some approach involving a crate which lives
separately, but I'm happy to be the "bumper" on this for now
- Relates to https://github.com/risc0/risc0/issues/441

---------

Co-authored-by: Erik Kaneda <erik@risczero.com>