docs: add Bonsai Eth examples (#1127)
Adding examples of how Bonsai and the zkVM can be used w/ Ethereum. 
I broke convention with the hierarchy of the zkVM examples doc and used
H3's for each example instead of a bulleted list because these examples
require additional context and more thorough descriptions. All of the
links in this additional context made the page difficult to read without
making this change.

---------

Co-authored-by: Brett Carter <brettcarter@Risczero-MBP.localdomain>
Co-authored-by: Frank Laub <flaub@risc0.com>