guest code: add `getrandom` feature to toggle getrandom in the guest (#1162)
Until now, the default behavior was to expose a `getrandom`
implementation to the guest that will call `sys_random` to generate
random numbers from the host. We did this because the `getrandom`
crate is a dependency of various crates and without exposing a
`getrandom` function, some crates may fail to compile for the zkvm.
While this is functionality allows more crates to compile, it could
unintentionally allow guest code depend on the host for random
numbers. Avoid this by gating `getrandom` function that uses
`sys_random` using a feature flag.

---------

Co-authored-by: Frank Laub <flaub@risc0.com>