Add the `resolve` recursion program (#1179)
This PR adds the `resolve` recursion program. With this program it
becomes possible to produce succinct proofs for zkVM executions that
utilize proof composition through the `env::verify` API for logically
verifying RISC Zero receipts within the RISC Zero zkVM.

As part of this PR, these is additional plumbing of `ReceiptMetadata`
structs such that the all required information is available for the
`resolve` program. In particular, the `seal` of a `SegmentReceipt`
includes the top level fields of the `ReceiptMetadata`, but does not
include the full `Output`. In order to run `resolve` the head of the
assumptions list must be available. So this PR adds a an explicit
`metadata` field to `SegmentReceipt` and ensures metadata is preserved
when passing through `lift` and `join` operations. The `pub(crate)
Merge` trait is added to `ReceiptMetadata` to help facilitate this and
also check for consistency through the steps of recursion. (This
actually identified a consistency issue with how the PC was being
handled in `risc0-zkvm` and in the recursion circuit.)

Resolves: https://github.com/risc0/risc0/issues/991

---------

Co-authored-by: Frank Laub <flaub@risc0.com>