Rework Bonsai on Ethereum docs (#1217)
Users are often confused about the current status of the Bonsai Relay.
In particular, users get confused as to whether they need to deploy the
relay themselves, and if there are alternatives to using it.

This PR reworks the Bonsai on Ethereum doc to clarify the current
status. Users must currently deploy them relay themselves. It also
provides a high-level description of how to use Bonsai and the verifier
contract directly, aiming to give enough context for an advanced
engineer/team to implement that flow.

This PR additionally deletes the example code that was embedded in this
doc. The example code appears to have been copied from the Foundry
template, but does not provide enough context to work independently from
the template. Instead, this PR has a link to the example in context.
This also helps mitigate potential version mismatch between the template
and the docs. We plan to overhaul the template soon, so this will be
important.

---------

Co-authored-by: Angelo Capossele <angelocapossele@gmail.com>
Co-authored-by: Paul <paul.gafni@gmail.com>