Fix entrypoint for the staticlib build of risc0-zkvm-platform (#1221)
Currently, the entrypoint defined in the `rust_rt.rs` file of
`risc0-zkvm-platform` passes a null pointer to the `sys_halt` function
to stop the program. When written, this was not detected as an error
because there existed a bug in the `Executor` and no usage every
attempted to prove a program linked with this entrypoint. The
`rust_rt.rs` file contains functionality to create a statically linkable
runtime for use in building programs that are not zkVM aware. Right now
the only usage of this is to build and run Rust tests with `cargo
risczero test`.

This PR fixes the entrypoint to pass a pointer to a constant array of
`[0u32; 8]`. It also refactors the entrypoint assembly to match the one
`risc0-zkvm::guest`, defining the `__start` symbol. It also refines the
definition of the `__start` symbol in both locations to encode that it
does not return. Returning from the `__start` symbol would cause
undefined behavior.

This PR does not de-duplicate these two definitions. We still have some
work to do on this front, but this PR does not seem like the time to
address this.