Update copyright to 2024 (#1287)
Happy New Year! I've omitted the circuit from this PR but I will make a
separate update after updating WIP and bootstrapping. It looks like tests in
doc-test/version-0.18 is failing due to the `risc0-zkvm-methods` crate
failing to build. I'm going to omit changes to that directory for the time being.