fix: remove unnecessary loads in ecall_sha (#1280)
Currently this will read past the first digest until the block is full
(2*digest). In the case that the second digest is directly after the
first in memory, this read is just redundant. In the case that they are
not, when reading just one, there are unnecessary reads done past the
first digest in memory.

I might be missing something here, but checking manually and running
some basic tests work, so just opening this to see if and what
changes/breaks to see if I am missing something.

---------

Co-authored-by: Frank Laub <flaub@risc0.com>