Distinguish between and avoid use of ExitCode::Fault and SessionLimit (#1317)
Currently our code does not distinguish between guest-reported/proven
and host-reported/unproven exit codes. This PR addresses this issue by
giving `ExitCode::SystemSplit`, `Fault` and `SessionLimit` distinct exit
codes.

The rv32im circuit does not support exiting with with a verified status
of `SessionLimit` or `Fault`. As a result, these exit codes should be
unused and reserved for future revision of the circuit. This PR
refactors the executor to reflect this and align with the fact that the
prover will never write these statuses to the seal.

Supersedes: #1313
Resolves: #1306

---------

Co-authored-by: Frank Laub <flaub@risc0.com>