Update Ethereum documentation to reflect new tenplate (#1469)
We just released a new version of our [Foundry
Template](https://github.com/risc0/bonsai-foundry-template) which
reflects an updated model of how to interact with Ethereum. Instead of
running a dedicated relay service, the model is to run a lighter weight
integration with a publisher component.

This PR updates the documentation to remove references to the relay from
0.20 and next API docs, to avoid confusion.

The relay model is still available, on tbe relay branch of the docs and
in the `risc0-ethereum` repo.