fix: bug with unaligned short read (#1557)
So what's currently happening is when requesting an unaligned amount of
bytes and the read is short, such that it crosses a word boundary, the
guest will overwrite the last word with the lastword return from the
host. This is a workaround to avoid changing the host.

Right now what this does is:

```
|----------request 15 bytes---------------|
|-----3x SyscallRecord words------|
00 01 02 03 04 05 06 07 08 09 0a 00
```

and then the last word writes over the end of the buffer, zeroing the
bytes written in the last word until aligned:

```
                        |lastword-|
00 01 02 03 04 05 06 07 00 00 00 00
```

Alternative being the host doesn't statically set the guest memory based
on `#requested.floor_div(4)`, but instead only create a `SyscallRecord`
with the full words written, and re-use this fill from last word logic.

```
|----------request 15 bytes---------------|
|--2x SyscallRecord---| |lastword|
00 01 02 03 04 05 06 07 08 09 0a
```

~~where currently what it does in this case (from changes in this PR) is
just removing writing the lastword if a short read past a word
boundary:~~ Edit: this functionality was changed to write zero bytes in
the lastword slot, such that zeroes are filled until the request length
```
|----------request 15 bytes---------------|
|-----3x SyscallRecord words------|
00 01 02 03 04 05 06 07 08 09 0a 00
```

Note that with this proposed implementation, as well as the current
implementation, if a lot more bytes are requested than available, there
are more words in the `SyscallRecord` created than necessary, setting
all of the buffer bytes to 0. Unclear to me how much of an impact this
has on proving.

```
|---------------request 19 bytes------------------------|
|----------4x SyscallRecord words-------------|
00 01 02 03 04 05 06 07 08 09 0a 00 00 00 00 00
```

I don't know what is more ideal, but hope this workaround or at least
one of these tests is helpful. Those tests were just created to help
narrow down the issue, happy to change them if this PR actually comes
through :)

---------

Co-authored-by: Victor Graf <victor@risczero.com>