Prover: Switch receipt kind based on `ProverOpts` (#1688)
This change adds a way to switch receipt formats using the `ProverOpts`.
There's a new field called `ReceiptFormat` that is used to specify the
desired receipt format generated from the prove function. Fake receipts
were not included as a part of this because `DevMode` has a separate
`ProverServer` implementation. This implements a way to configure
`Prover::prove` to return either composite or succinct receipts. In the
near future, an implementation for `CompactReceipts` will be implemented
for x86-linux as well as a way to configure the receipt types for
`Prover::Compress`.

Closes: #1689

---------

Co-authored-by: Frank Laub <flaub@risc0.com>