`ProverImpl::prove_session`: enable compact receipts (#1717)
This change adds a case to enable compact receipts using the `Prover`
trait. The stark2snark function being called ensures that the host is
running on x86 and that docker is installed. If the docker image does
not exist locally, the first invocation of this command will
automatically download the image from docker hub so there is no need to
check ahead of time if the image exists.

---------

Co-authored-by: Frank Laub <flaub@risc0.com>