[security model doc] translate from cost estimate to time estimate (#1664)
This doc updates the description for "How secure is the STARK protocol?"
from a cost-based estimate to a time-based estimate.

The primary motivation to update this is that the current estimate is
dependent on a few assumptions from 2020. Switching from cost to time
removes the dependency on both bitcoin hashrate and price of bitcoin.

The draft here was written by Kevin; I'm hesitating on the specifics
because it seems weird to be referencing `sha1`. Would Poseidon2 be the
appropriate benchmark for this?