Support producing `CompactReceipt` through `BonsaiProver` and improve `Prover::compress` (#1762)
This PR adds support for producing a `CompactReceipt` through the
`BonsaiProver` impl of the `Prover` trait. This unblocks moving proving
in `risc0-ethereum` and the Foundry template to the `risc0-zkvm` API
(https://github.com/risc0/risc0-foundry-template/issues/92), and helps
simplify the implementation of receipt metadata (coming in a PR soon).

This PR additionally fixes up some missing logic in the
`Prover::compress` implementations, and changes the semantics of
`opts.receipt_kind`. With this PR, the receipt kind field is interpreted
a _minimum_ compression level, rather than exact. As a result, it is no
longer an error to call `compress` on a `Compact` receipt, with
`ReceiptKind::Succinct`. Instead, it is a no-op. I believe this better
aligns with the expectation of a caller, and prevents needing to check
the receipt type before calling compress.

---------

Co-authored-by: Erik Kaneda <erik@risczero.com>
Co-authored-by: Frank Laub <flaub@risc0.com>