cargo-risczero: drop cargo-generate from new command (#1867)
The `cargo-generate` crate is designed with the assumption that
templates to reside in a small repository. Our template is a part of a
large repository and takes anywhere between 3-10 seconds to run `cargo
risczero new` because it has to clone the repo. This change drops
cargo-generate and instead uses regex patterns to search and replace the
template during generation. Locally, this takes about 0.5 seconds to
run. I've dropped several options that seemed irrelevant. Another
notable change is the renaming of the template's `Cargo.toml` to
`Cargo-toml`. By doing so, I have fixed
https://github.com/risc0/risc0/issues/1840.

Fixes: #1840