Fix Website CI Triggering `bench-status-check` (#1879)
Using GitHub's built-in `paths:` within `on:` conditions in a workflow
with a job using `if: always()` causes the job to always be triggered -
regardless of the paths referenced. This does not make sense logically,
but was proved via testing.

When this happens, the job registers as Required but does not trigger
because the jobs it depends on to run are not triggered and ran. As a
result, the CI hangs and changes cannot be merged since a "Required" job
has not succeeded:

<img width="600" alt="image"
src="https://github.com/risc0/risc0/assets/133707571/db36dd8c-d50e-448d-8914-0cd80f17c4f8">

`paths:` was added to `bench_pr.yml` last week here:
https://github.com/risc0/risc0/pull/1856/commits/d876ce24783dd28ecfad429a594e5fc92a5d6877

This issue did not present itself until now since other PRs after the
change included changes that triggered the jobs in `bench_pr.yml`
([example](https://github.com/risc0/risc0/pull/1856/files)).

---

**Fix:**

- [x] Add a space to file under `website` to only trigger Website CI,
but verify that it still triggered `bench-status-check` when it
shouldn't have
  - [ ] This can be reverted back before merging
- [x] Replace `paths:` in `on:` conditions with a `changes` job that
uses the `paths-filter` action
- [x] Test that all CI checks pass after change