bump version `v1.0.0-rc.5` to `v1.0.0-rc.6` (#1904)
This PR bumps the version of the `risc0` crates to `1.0.0-rc.6` and
bumps the `bonsai-sdk` to `rc.3`. This change has the usual Cargo.toml,
Cargo.lock, and image ID changes. In addition, I also noticed that
`cargo test -p risc0-zkvm` failed to compile. We build and test with the
`-F prove` but noticed that the risc0-zkvm tests fail to build without
this flag. I've added an extra test step to ensure that this path is
being tested and added a few `cfg` statements to fix the build error.