Add new function to Groth16Receipt (#1905)
This PR added a `Groth16Receipt::new` function, allowing users of the
crate to construct the `Groth16Receipt` type if they have the
constituent parts. This is being added to support users of the Bonsai
REST API, as this API returns the parts of the receipt, from which the
caller needs to construct the `Groth16Receipt` object.