Re-enable website check CI job (#1957)
It appears that the `check` CI job was disabled due to a
misconfiguration in the `website.yml` config file. Unfortunately, this
has been a silent failure and so I only noticed it when I happened to
run `yarn fmt` locally. This PR adds the missing `needs: changes` line
to fix the job.