Automatically set the branch for `risc0-ethereum` (#2066)
As currently `risc0-ethereum` main is (and should be) consistent with
`risc0` main, whenever we create a back port PR to some other branch, we
might see the risc0-ethereum CI failing. To solve that, this PR
instructs CI to conditionally pick the branch of `risc0-ethereum` based
on the name of the branch of `risc0` the PR is being merged into.