fix: add n_bytes_read check for C example inputs (#2060)
Noticed when [doing something unrelated that re-used some of this
code](https://github.com/austinabell/picoprogram/pull/1) that this API
is a bit misleading as it ignores the number of bytes read and doesn't
allow checking explicitly. Given short reads are allowed, this will just
interpret the values as zero, which could be a footgun.

Also added the check to avoid the ambiguity of not inputting enough
bytes for the values, as it's probably best to recommend people to do
something similar. Alternatively, could rename the function to
`env_read_exact` and do this assertion on the Rust side, I don't have a
preference.