client/server: improve version mismatch error (#2123)
This change implements version checking on the client side. Upon
creation of the ParentProcessConnection, the client will query the r0vm
binary for its version and emit a verbose error message upon a version
mismatch. We have previously discussed that this could be a handshake
bewteen the client and server but I saw a downside to this solution: if
we change the server to get this to work using `Client.connect`, it
means that this solution will not work for older servers and will only
work for new servers. For example, if a user decides to use an older
server (i.e. v0.21) on a newer client, we will end up getting the old
error message which was difficult to understand:

```
thread 'main' panicked at risc0/r0vm/src/lib.rs:192:18:
called `Result::unwrap()` on an `Err` value: incompatible client version: 1.1.0-alpha.1, server version: 1.0.3
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
thread 'main' panicked at host/src/main.rs:41:10:
called `Result::unwrap()` on an `Err` value: failed to fill whole buffer
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
```

Improving client/server compatibility based off of changing both client
and server will only work for new servers. Where as changing only the
client will allow the version check to work for older servers.

This implementation will result in an error during proving or execution.
We discussed implementing this check as a part of `default_prover()` but
this function returns a trait object. Since `default_prover` is
infallable, including the compatibility check as a part of this function
will either force panics or change the type signature of this function.
Therefore, I went with adding this check at the time of the subprocess
connection which happens at the beginning of execution.

---------

Co-authored-by: Frank Laub <frank@risczero.com>
Co-authored-by: Frank Laub <flaub@risc0.com>