Plumb ReceiptClaim through SessionInfo and client-server (#2342)
This PR plumbs the `session.claim()` through to the `SessionInfo` such
that it is available from executions carried out with the `Executor`
trait. Motivation for this PR is that I am writing some tests of
composition and want to be able to run the executor and convert the
result into a fake receipt. This is possible with `ExecutorImpl`, but
not with `Executor` right now.

More context on
https://github.com/risc0/risc0/issues/2267#issuecomment-2347372529

Closes #2267