Add .watchmanconfig to fix editor integration with neovim and CoC (#2428)
I use neovim and CoC for my editor, which uses watchman on MacOS.
Watchman has started to die on me when it tries to process the repo. I
found that it needs to ignore the target directories or else it falls
over since the repo has become quite large.

I don't feel strongly that this gets merged. I can also maintain it
locally, but wanted to raise this in case there is anyone else out there
that might hit the same issue.