Add RSA syscall (#2481)
Extends our RSA acceleration support to enable accelerated computation
of the RSA exponentiation (instead of just accelerated verification of
the result of an RSA exponentiation). This is accomplished by adding an
RSA syscall which allows the guest to request that the host provide the
result of the RSA operation, which it can then verify via the recursion
circuit bigint RSA accelerator.

This PR also adds a convenient API for accessing this RSA acceleration:
`modpow_65537`. This takes in two `BigUint`s, a `base` and a `modulus`,
and computes `base^65537 % modulus` (where `^` is exponentiation) via
this syscall + bigint proof mechanism. For convenience in patching into
other crates, two bigint libraries are available for this API: By
default, it will use `num_bigint`, but if the `bigint-dig-shim` feature
of `risc0-circuit-bigint` is enabled it will instead use
`num_bigint_dig`.

---------

Co-authored-by: Frank Laub <flaub@risc0.com>