bump version to 1.3.0-alpha.1 (#2547)
This PR bumps crate versions to `1.3.0-alpha.1`. In doing the 1.2.0-rc.1
release, I noticed that bonsai-sdk was not set to use the version
identified in the workspace. This something that I forgot to backport
from release-1.1. When I ran `cargo publish release -x` it warned me
about a few missing descriptions and it was trying to publish `rzup`.
This new behavior that I hadn't observed in the past but I've added
descriptions to the missing crates and disabled `rzup` from being
published.