Add used and free methods to bump allocator module (#2677)
This PR adds `used` and `free` methods to the bump allocator module to
allow for collecting memory metrics. It also moves the logic of the bump
allocator into the `heap::bump` module, ensuring that it cannot be used
if it is not enabled at compile time (which would cause undefined
behavior).

Also, this PR adds the `unsafe` specifier to `heap::embedded::init`.
This should be specified because calling this function can result in
breaking Rust's memory guarentees. This is a breaking change, however
any user code that calls this function (if any exists) is likely
unsafe/broken right now, since init is called as part of the guest
entrypoint and should not be called again.

---------

Co-authored-by: Austin Abell <austinabell8@gmail.com>