ZKVM-943: Require addresses used by `ecall_sha`, `ecall_bigint`, and `ecall_bigint2` to be aligned in emulator (#2727)
Currently there is an issue in that a program can issue an `ecall`, e.g.
for bigint (1), with an unaligned pointer and the executor will succeed,
but proving will fail. This is an issue both for the smaller impact that
it's better to know sooner when an execution cannot be proven, and for
the larger impact that it could be leveraged to cause issues on
Boundless, where the executor is used to determine whether or not a
given job is provable.

This PR also adds a check to the software ecall that the _end_ of the
guest buffer is in guest memory. This is related check prevents the host
from potentially allocating up to 4 GB of memory when the given buffer
len is actually invalid.

---------

Co-authored-by: Erik Kaneda <erik@risczero.com>
Co-authored-by: Frank Laub <flaub@risc0.com>