ZKVM-975: Remove default timeout (30s) and replace with None timeout (#2743)
Closes #2742 

This little fix does the trick. By default the `HttpClient` has a 30s
timeout and this is getting triggered for my use-case. Explicitly
setting this to None removes the timeout and solves the issue.

It may be the case that some timeout is desired and if this is the case
we could introduce another env var to set this.