ZKVM-1035: Improvements to guest profiling (#2801)
This adds two main improvements to guest profiling
- Support for inline functions
- Show page events as function calls

### inline functions
To maintain the way the cycle-accurate profiler works we build a table
of all the inline function from the DWARF data and then when we enter an
inline function we push a frame to the profiler's call stack for that
inline function. Once we leave the range of the inline function we pop
it off.

### page events
When we incur something that causes a page to be loaded we now magic up
a functions called `[PageIn]` and `[PageOut]`.
Since this can happen somewhat randomly within a segment, functions that
normally would be seen as low-cost can incur a high cycle count due to
them triggering page events (you can see this in the profile below with
`FdReader::new`, that function is only a single instruction which writes
a single word to memory)
The emitting of these events was added to the v1 circuit code. The v2
circuit code needs to be updated to work with the profiler first, and
that isn't tackled in this PR.

### comparison
These screenshots are from before and after the changes running
`tests::keccak_update2::case_1`

before:
<img width="1591" alt="before"
src="https://github.com/user-attachments/assets/9b3a4f2c-8dfe-46eb-bc3e-932088e554ab"
/>
after:
<img width="1584" alt="after"
src="https://github.com/user-attachments/assets/f44b408e-ff30-42b0-9945-a5de0fc297f7"
/>