ZKVM-1224: Add ABI version number to ProgramBinary; Check version in zkVM (#2971)
This adds a new header to the ProgramBinary. It stores it on disk as a
list of enum values encoded using `postcard`. I wanted to use `bincode`
at first, but it doesn't support `no_std` usage. It ignores values that
it doesn't know about, potentially allowing for future versions of the
header to still be decoded even if they contain new kinds of header
values.

The value of this version is checked in the zkVM and an error is thrown
if it doesn't match the one value we currently know about. This will
prevent version 2.0.0 of the zkVM from attempting to execute guest
programs built by newer version of the code. Also, this will allow
future versions of zkVM to detect older versions of guest programs and
handle that appropriately.