ZKVM-1260: Account for missing ControlDone cycle in executor (#3015)
The `segment_cycles()` function in the executor accounts for the number
of cycles that are used so far in the segment. Its calculation was
missing the extra `ControlDone` cycle preflight adds (off by one.) This
mistake didn't show up when splitting because there was another bug in
the splitting logic where we used `>=` instead of `>`, (this negated the
extra `+1`). Later though, when we calculate the po2 for the final
segment we use the `segment_cycles()` function again, but this time the
off by one issue matters and we calculate the wrong po2 for the segment.
This then results in a crash later in preflight.

I'd like to create a guest program that manages to hit this bug without
the fix that we can add as a test, but its a bit tricky to accomplish so
I wanted to get this out there first without it.