ZKVM-1273: Widen file-lock around rzup install command (#3027)
`rzup` didn't really handle concurrent `install` commands being issued
on the same machine. I'm not sure I believe that it ever did really.

It did have a file-lock used when downloading a file, but this wasn't
sufficient since it didn't hold the lock long enough. Replace this file
locking with a new lock that is taken for the duration of installation.

Also, we can't delete the lock-file when we are done with it, because
you might delete a lock-file which another process has locked.