ZKVM-1266: Drop `risc0-foundry-template` from `risc0-ethereum` CI (#3020)
This PR removes the `risc0-foundry-template` from `risc0-ethereum` CI.
The purpose of this CI test should be to check if a given release (or
branch) of the risc0 repo breaks a public facing version of consumer
projects, `risc0-ethreum` and the `risc0-foundry-template` specifically.
However, the latter cannot be updated properly before the former gets
also updated.

Co-authored-by: Remi Bernotavicius <bobbobbio@gmail.com>