ZKVM-1308: Fix & run Clippy for risc0-bigint2 (#3070)
This adds a run with `-F unstable` to the clippy CI job. It also adds
the corresponding fixes to risc0-bigint2. This caught a number of
overlooked issues that only affected `unstable` functionality.

This includes the commit from https://github.com/risc0/risc0/pull/3069,
which is a change clippy would have caught.

---------

Co-authored-by: Barajeel <0xsilomir@gmail.com>