ZKVM-1314: Run fmt-markdown action on docs (#3084)
When working on https://github.com/risc0/risc0/pull/3083, I realized
that running `bun run fmt-markdown` in the `website` directory results
in changes (e.g. removed trailing whitespace). It seems that the `lint`
action is not cactching this kind of lack of formatting. One issue would
appear to be that the `lint` action does not use the same plugins, but
this does not actually fix this issue either.

This PR commits the results of `bun run fmt-markdown`. It fixes one
issue that might result in missed linting in our docs. It does not
address the issue of failing to check strict formatting.

---------

Co-authored-by: capossele <angelo@risczero.com>