ZKVM-1310: semver-checks: Implement checking of crate dependency version requirements (#3074)
We had an issue where we ended up breaking the 1.2.X build of the risc0
crates when we published 1.4.X versions of crates as part of the 2.0.0
release. This was because the 1.4.X crates ended up depending on 2.0.X
versions of crates (`risc0-core` and `risc0-zkp` at a minimum) which
conflicted with dependencies in other 1.2.X crates. This caused cargo to
include both versions. This can be okay, but in this case it causes
issues where the types are incompatible (because there are two different
versions of these crates).

This adds a check to semver-checks that sees if a crate has a workspace
dependency version requirement increase in major version, and if it
does, we require that crate to also bump its major version. In the above
situation it should have forced the 1.4.X crates to bump their versions
to 2.0.X.