ZKVM-1400: Executor Bigint Smallvec Spillage (#3189)
In the executor where bigints are handled, when loading, a Smallvec is
allocated to read limbs into. Currently the inlined smallvec value is
optimized for 8 u32 limbs. Although this code isn't related to BLS, BLS
is used often by many people in the ZKVM and limbs sizes (at least in
blst are 12 u32s), so through some profiling, I observed that we are
constantly spilling over and heap allocating here anyways.