ZKVM-1020: Fix verifier parameters metadata check to actually compare verifier context and metadata (#2619)
Discovered while debugging a different issue, the check that is supposed
to report if verifier parameters are mismatched currently does not.
Instead it just checks internal consistentency of
`receipt.metadata.verifier_parameters` against
`receipt.inner.verifier_parameters`, both of which are digests to help
the verifier determine which version of the verifier parameters it
should use. Note that the digest alone is not enough to verify the
receipt, and instead the actual parameters come from the
`VerifierContext`.

This PR fixes the check to actually compare against the verifier context
so that the correct error message can be shown, instead of a generic
verification failure.

---------

Co-authored-by: Erik Kaneda <erik@risczero.com>
Co-authored-by: Frank Laub <flaub@risc0.com>
Co-authored-by: Tim Zerrell <tim.zerrell@risczero.com>